requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
sentence-transformers==2.7.0
faiss-cpu==1.7.4
rank-bm25==0.2.2
scikit-learn==1.3.2
beautifulsoup4==4.12.2
readability-lxml==0.8.1
aiohttp==3.9.1
tabulate==0.9.0
openpyxl==3.1.2
celery[redis]==5.3.6
google-api-python-client==2.109.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
tqdm==4.66.1
prometheus-client==0.19.0
redis==4.6.0
unidecode==1.3.7
lxml==4.9.3
requests==2.31.0
numpy==1.24.4
httpx==0.25.2
Jinja2==3.1.2
psutil==7.0.0